refactor(db): split commandLocks.ts to reduce file complexity
commandLocks.ts was over the qlty per-file complexity limit. Extracted
into three focused modules:

- commandStringUtils.ts (new, 61 lines): SqlExecutor type, string
  normalisation helpers, and error types (CommandNotFoundError,
  CommandConflictError, isMysqlDuplicateEntryError)
- commandConflicts.ts (new): all conflict-assertion and user-assignment
  functions including the deadlock-retry loop in
  assignUserToCommandWithinTransaction; imports isDeadlockError /
  MAX_DEADLOCK_RETRIES from commandLocks to avoid circular deps
- commandLocks.ts: named-lock primitives, exists-check queries,
  runSerializedCommandWrite (with deadlock retry), isDeadlockError,
  and MAX_DEADLOCK_RETRIES

commandLocks.ts re-exports all types moved to commandStringUtils so
db.ts and barrel callers are unaffected. Direct importers (counters.ts,
customCommands.ts) updated to import from the appropriate new modules.
